# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberSentinel API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/sessions:
    get:
      operationId: listSessions
      tags: [chat]
      summary: List chat sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [chat]
      summary: Create chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /chat/sessions/{id}:
    delete:
      operationId: deleteSession
      tags: [chat]
      summary: Delete chat session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /chat/sessions/{id}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send message to session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
...
    Session:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
    SessionInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    Message:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - role
        - content
    MessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content