# .migration-backup/lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CyberSentinel API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  role: string;
  content: string;
  createdAt?: string;
}